profile feature

+++ service/src/scripts/compose-dev.sh
@@ -3,7 +3,7 @@
 # move to src dir
 cd ../
 
-if [ $1 = "down" ]
+if [ ! -z $1 ] && [ $1 = "down" ]
 then
     docker-compose -f docker-compose.dev.yaml -p oop-infra down
 else
